completed authentication

+++ frontend/src/components/WorkoutDetails.js
@@ -1,15 +1,23 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
 import formatDistanceToNow from "date-fns/esm/formatDistanceToNow";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({workout}) => {
 
     const { dispatch } = useWorkoutsContext();
-
+    const { user } = useAuthContext()
     const handleClick = async () => {
+        if(!user){
+            return
+        }
+
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers:{
+                'Authorization': `Bearer ${user.token}`
+              }
         })
-        const json = response.json()
+        const json = await response.json()
         if(response.ok){
             dispatch({type: 'DELETE_WORKOUT', payload: json})
         }
